Fixes typo in alternative implementation of constructor. (problem found by Jan Friso)
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@966 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ src/squadt/utility/relocator.cpp
@@ -18,7 +18,7 @@ class ToolExecutor;
 class Specification;
 
 /* Stubs */
-ToolExecutor::ToolExecutor(unsigned int) {
+ToolExecutor::ToolExecutor(size_t) {
 }
 
 ToolExecutor::~ToolExecutor() {
